WIP refactoring Reacting JavaScript
THese things should get a JS context and talk to it. The ThreadWorker
is a detail of how the communication works and shouldnt be expose to
the world.

Currently Source/Fuse.Testing/TestRootPanel.uno still needs the Worker
to be internally accesible. We will fix this asap.

Next priority is spliting the Mirror and Casting functions out of the
threadworker. They can bee their own helper objects and live on the
Context