ThreadWorker.FuseJS is now only available on the JS Thread
This means that all of ThreadWorkers public interface except
`CheckAndThrow()` should be threadsafe.

Due to the change in how we construct (and that we shared contexts
across thread) we only need to ensure that only 1 Context is being
created at a time, so we change the `_ready` signal for a mutex.