Add cancelPendingNotifications to Fuse.LocalNotifications
This feature allows you to cancel all local notifications sent by the
app which are enqueued but not yet delivered.

On iOS the implementation was simple due to the built in function for
this.

On Android we built on top of a StackOverflow post which keeps
track off the LocalNotification IDs in android SharedPreferences.
By storing the IDs we can recreate the PendingIntents and cancel
them.

This did require rewriting `Later` on android to use this utils